Add experimental notice to "Search" toggle
